+++ tsconfig.json
@@ -13,7 +13,7 @@
         // "checkJs": true,                       /* Report errors in .js files. */
         // "jsx": "preserve",                     /* Specify JSX code generation: 'preserve', 'react-native', or 'react'. */
         "declaration": true /* Generates corresponding '.d.ts' file. */,
-        // "declarationMap": true,                /* Generates a sourcemap for each corresponding '.d.ts' file. */
+        "declarationMap": true /* Generates a sourcemap for each corresponding '.d.ts' file. */,
         "sourceMap": true /* Generates corresponding '.map' file. */,
         // "outFile": "./",                       /* Concatenate and emit output to single file. */
         "outDir": "dist" /* Redirect output structure to the directory. */,
@@ -61,5 +61,6 @@
         // "experimentalDecorators": true,        /* Enables experimental support for ES7 decorators. */
         // "emitDecoratorMetadata": true,         /* Enables experimental support for emitting type metadata for decorators. */
     },
+    "include": ["src/**/*.ts", "src/**/*.d.ts"],
     "exclude": ["node_modules", "test", "types"]
 }
